create separate constants file
