+++ src/lib/constants.json
@@ -1,0 +1,8 @@
+{
+  "sevenDays": 6.048e+8,
+  "accessTokenUrl": "https://www.paypal.com/muse/api/partner-token",
+  "storage": {
+    "paypalCrCart": "paypal-cr-cart",
+    "paypalCrCartExpiry": "paypal-cr-cart-expiry"
+  }
+}
