+++ src/tracker-component.js
@@ -8,7 +8,11 @@ import generate from './generate-id';
 import { getCookie, setCookie } from './lib/cookie-utils';
 import getJetlore from './lib/jetlore';
 import { getDeviceInfo } from './lib/get-device-info';
-import { removeFromCart, addToCart } from './lib/compose-cart';
+import { composeCart } from './lib/compose-cart';
+import {
+    accessTokenUrl,
+    storage
+} from './lib/constants';
 import type {
     TrackingType,
     CartEventType,
@@ -23,15 +27,6 @@ import type {
     Config
 } from './lib/types';
 
-const storage = {
-    paypalCrCart: 'paypal-cr-cart',
-    paypalCrCartExpiry: 'paypal-cr-cart-expiry'
-};
-
-const sevenDays = 6.048e+8;
-
-const accessTokenUrl = 'https://www.paypal.com/muse/api/partner-token';
-
 const getUserIdCookie = () : ?string => {
     return getCookie('paypal-user-id') || null;
 };
@@ -39,42 +34,6 @@ const getUserIdCookie = () : ?string => {
 const setRandomUserIdCookie = () : void => {
     const ONE_MONTH_IN_MILLISECONDS = 30 * 24 * 60 * 60 * 1000;
     setCookie('paypal-user-id', generate.generateId(), ONE_MONTH_IN_MILLISECONDS);
-};
-
-const composeCart = (type, data) => {
-    // Copy the data so we don't modify it outside the scope of this method.
-    let _data = { ...data };
-
-    // Devnote: Checking for cookie for backwards compatibility (the cookie check can be removed
-    // a couple weeks after deploy because any cart cookie storage will be moved to localStorage
-    // in this function).
-    const storedCart = window.localStorage.getItem(storage.paypalCrCart) || getCookie(storage.paypalCrCart) || '{}';
-    const expiry = window.localStorage.getItem(storage.paypalCrCartExpiry);
-    const cart = JSON.parse(storedCart);
-    const currentItems = cart ? cart.items : [];
-
-    if (!expiry) {
-        window.localStorage.setItem(storage.paypalCrCartExpiry, Date.now() + sevenDays);
-    }
-
-    switch (type) {
-    case 'add':
-        _data.items = addToCart(data.items, currentItems);
-        break;
-    case 'set':
-        _data.items = data.items;
-        break;
-    case 'remove':
-        _data = { ...cart, ...data };
-        _data.items = removeFromCart(data.items, currentItems);
-        break;
-    default:
-        throw new Error('invalid cart action');
-    }
-
-    window.localStorage.setItem(storage.paypalCrCart, JSON.stringify(_data));
-
-    return _data;
 };
 
 const getAccessToken = (url : string, mrid : string) : Promise<Object> => {
