+++ test/tracker-component.test.js
@@ -746,7 +746,7 @@ describe('paypal.Tracker', () => {
         const email = '__test__email3@gmail.com';
         const userName = '__test__userName3';
         // clear local storage to ensure a request happens
-        window.localStorage.removeItem('property-id-abcxyz123-xyz')
+        window.localStorage.removeItem('property-id-abcxyz123-xyz');
         Tracker({ user: { email, name: userName } });
 
         expect(appendChildCalls).toBe(0);
@@ -758,7 +758,7 @@ describe('paypal.Tracker', () => {
         const email = '__test__email3@gmail.com';
         const userName = '__test__userName3';
         // clear local storage to ensure a request happens
-        window.localStorage.removeItem('property-id-abcxyz123-xyz')
+        window.localStorage.removeItem('property-id-abcxyz123-xyz');
 
         Tracker({ user: { email, name: userName }, propertyId: 'hello' });
         expect(fetchCalls.length).toBe(0);
